requirements.txt
click  == 8.*
pandas == 2.0.*
requests == 2.*
python-dotenv == 1.*
jsonschema == 4.*
